requirements_minimal.txt
# Minimal Requirements for DentalAI - CPU Server Only
# این فایل فقط dependency های ضروری برای سرور CPU را شامل می‌شود

# ============================================================================
# Core API Framework (فقط Flask - FastAPI استفاده نمی‌شود)
# ============================================================================
flask>=2.0.0,<2.3.0
flask-cors>=4.0.0

# ============================================================================
# Image Processing (ضروری)
# ============================================================================
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ============================================================================
# Deep Learning Framework (ضروری برای مدل‌های AI)
# ============================================================================
# PyTorch CPU-only version (حجم کمتر از نسخه کامل)
# ⚠️ مهم: از نسخه CPU-only استفاده کنید تا حجم کمتری داشته باشد
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0

# ============================================================================
# Intra-Oral Analysis (YOLO) - ضروری
# ============================================================================
ultralytics>=8.0.0

# ============================================================================
# Facial Landmark Detection - ضروری
# ============================================================================
mediapipe>=0.10.0

# ============================================================================
# CLdetection2023 Model Requirements - ضروری
# ============================================================================
# OpenMMLab tools for MMPose
openmim
mmengine>=0.6.0,<1.0.0
mmcv>=2.0.0rc4,<=2.1.0

# ============================================================================
# Additional utilities (فقط اگر استفاده می‌شود)
# ============================================================================
# scipy فقط برای LAB model استفاده می‌شود (اگر LAB استفاده نمی‌کنید، حذف کنید)
scipy>=1.7.0

# ============================================================================
# Dependency های حذف شده (استفاده نمی‌شوند):
# ============================================================================
# ❌ fastapi - استفاده نمی‌شود (فقط Flask استفاده می‌شود)
# ❌ uvicorn - استفاده نمی‌شود (فقط Flask استفاده می‌شود)
# ❌ python-multipart - استفاده نمی‌شود (فقط Flask استفاده می‌شود)
# ❌ scikit-image - استفاده نمی‌شود
# ❌ python-dateutil - استفاده نمی‌شود
# ❌ dlib - استفاده نمی‌شود (فقط mediapipe استفاده می‌شود)
# ❌ face-alignment - استفاده نمی‌شود
# ❌ retina-face - استفاده نمی‌شود



